requirements.txt
py-clob-client
python-dotenv
requests
websocket-client
